Refactored component files to includes CSS
