--- src/components/Board/constants.js
@@ -1,6 +1,0 @@
-
-const NUM_ROWS = 6;
-const NUM_COLUMNS = 7;
-
-export { NUM_ROWS, NUM_COLUMNS };
-
