+++ src/components/Cell/Cell.css
@@ -1,0 +1,4 @@
+.Cell {
+    width: 25px;
+    height: 25px;
+}
