+++ src/components/Cell/Cell.js
@@ -1,0 +1,10 @@
+import './Cell.css';
+export const Cell = (props) => {
+
+  return (
+    <td className={"Cell"} onClick={() => props.handleClick(props.rowIdx, props.colIdx)}
+        style={{backgroundColor: props.cell.color}}>
+    </td>
+  );
+}
+
