--- src/components/Cell/CellIndex.js
@@ -1,9 +1,0 @@
-export const Cell = (props) => {
-
-  return (
-    <td onClick={() => props.handleClick(props.rowIdx, props.colIdx)} width="50px" height="50px"
-        style={{backgroundColor: props.cell.color}}>
-    </td>
-  );
-}
-
